Refactor backend API and adding bot function
