--- service/app/views/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-blue_channel = Blueprint('channel', __name__)
-
-from . import channel
