+++ src/main.py
@@ -1,0 +1,6 @@
+from robot.video_bot import most_popular_job
+from crawler import videos
+from storage import video_storage
+if __name__ == '__main__':
+
+    most_popular_job("TW")
