+++ src/oauth/api_oauth.py
@@ -1,0 +1,9 @@
+import googleapiclient.discovery
+
+from utils.config import API_DEVELOPER_KEY
+
+api_service_name = "youtube"
+api_version = "v3"
+
+youtube = googleapiclient.discovery.build(
+    api_service_name, api_version, developerKey=API_DEVELOPER_KEY)
