+++ src/views/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+blue_channel = Blueprint('channel', __name__)
+
+from . import server
